Destroy all caches to work with angular-mocks tests
angular-cache should clean itself to work with angular unit tests.
angular-mocks calls `$rootScope.$destroy` in its cleanup phase for every
spec, so angular-cache can clean itself in an event listener on the
`$destroy` event.

Without this, angular-cache errors when more than one spec injects the
CacheFactory and the app code is creating caches with it.